Move ToggleUser and add isDeleted for a user
