+++ src/components/Admin/Authentication/SignOut.tsx
@@ -1,5 +1,5 @@
-import React, {useState} from 'react';
-import {SubmitButton} from '../ComponentUtil';
+import React from 'react';
+import {Button, Box} from '@chakra-ui/react';
 import {firebaseAuth} from '../../../firebase/firebase';
 import {useTranslation} from 'react-i18next';
 
@@ -7,31 +7,26 @@ import {useTranslation} from 'react-i18next';
  * Button that signs the user out
  */
 const SignOut: () => JSX.Element = () => {
-  const [error, setError] = useState('');
   const {t} = useTranslation('administration');
   /**
    * Signs out the user and sets authentication status to false.
    * @param event - submit form event
    */
-  async function handleSignOut(event: React.FormEvent<HTMLFormElement>) {
-    // Prevents submission before sign out is complete
-    event.preventDefault();
-
-    try {
-      await firebaseAuth.signOut();
-    } catch {
-      setError(t('signOutError'));
-    }
+   function handleSignOut(): Promise<void> {
+    return firebaseAuth.signOut();
   }
 
   return (
-    <form onSubmit={handleSignOut}>
-      <SubmitButton
-        label={t('signOut')}
-        color={'red'}
-        error={error}
-      ></SubmitButton>
-    </form>
+    <Box justifyContent="center">
+      <Button
+        minWidth="50%"
+        marginY={4}
+        onClick={handleSignOut}
+        colorScheme="red"
+        >
+        {t('signOut')}
+      </Button>
+    </Box>
   );
 };
