+++ src/contexts/AuthContext.tsx
@@ -10,6 +10,7 @@ import {Props} from './AppProviders';
  * - `isAdmin` if user is an admin
  * - `name` user's name or empty string
  * - `email` user's email
+ * - `isDeleted` if a user's account is scheduled for deletion
  */
 interface AuthInterface {
   isAuthenticated: boolean;
@@ -17,6 +18,7 @@ interface AuthInterface {
   isAdmin: boolean;
   name: string;
   email: string;
+  isDeleted: boolean;
 }
 
 /**
@@ -36,6 +38,7 @@ const AuthProvider: React.FC<Props> = ({children}: Props) => {
   const [isLoading, setIsLoading] = useState(false);
   const [email, setEmail] = useState('');
   const [name, setName] = useState('');
+  const [isDeleted, setIsDeleted] = useState(false);
   // --------------- End state maintenance variables ------------------------
 
   /**
@@ -45,6 +48,7 @@ const AuthProvider: React.FC<Props> = ({children}: Props) => {
     setIsAdmin(false);
     setEmail('');
     setName('');
+    setIsDeleted(false);
   }
 
   useEffect(() => {
@@ -81,10 +85,14 @@ const AuthProvider: React.FC<Props> = ({children}: Props) => {
             const userData = snapshot.data();
 
             if (userData) {
-              if (typeof userData.admin === 'boolean')
+              if (typeof userData.admin === 'boolean') {
                 setIsAdmin(userData.admin);
+              }
               if (typeof userData.name === 'string') setName(userData.name);
               if (typeof userData.email === 'string') setEmail(userData.email);
+              if (typeof userData.isDeleted === 'boolean') {
+                setIsDeleted(userData.isDeleted);
+              }
             }
             setIsLoading(false);
           } else {
@@ -94,6 +102,7 @@ const AuthProvider: React.FC<Props> = ({children}: Props) => {
               name: user.displayName ?? '',
               email: user.email ?? '',
               admin: false,
+              isDeleted: false,
             };
             firestore
               .collection('users')
@@ -119,6 +128,7 @@ const AuthProvider: React.FC<Props> = ({children}: Props) => {
         isLoading: isLoading,
         name: name,
         email: email,
+        isDeleted: isDeleted,
       }}
     >
       {children}
@@ -128,7 +138,7 @@ const AuthProvider: React.FC<Props> = ({children}: Props) => {
 
 /**
  * Custom hook to allow other components to use authentication status
- * @returns `{isAuthenticated, isLoading, isAdmin, name, email}`
+ * @returns `{isAuthenticated, isLoading, isAdmin, name, email, isDeleted}`
  */
 const useAuth: () => AuthInterface = () => useContext(AuthContext);
